web - separate factory from API
