+++ src/vs/workbench/browser/web.api.ts
@@ -3,44 +3,37 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { IWorkbench, main } from 'vs/workbench/browser/web.main';
 import { UriComponents, URI } from 'vs/base/common/uri';
-import { IWebSocketFactory, IWebSocket } from 'vs/platform/remote/browser/browserSocketFactory';
+import { IWebSocketFactory } from 'vs/platform/remote/browser/browserSocketFactory';
 import { IURLCallbackProvider } from 'vs/workbench/services/url/browser/urlService';
 import { LogLevel } from 'vs/platform/log/common/log';
-import { IUpdateProvider, IUpdate } from 'vs/workbench/services/update/browser/updateService';
-import { Event, Emitter } from 'vs/base/common/event';
-import { Disposable, IDisposable, toDisposable } from 'vs/base/common/lifecycle';
-import { IWorkspaceProvider, IWorkspace } from 'vs/workbench/services/host/browser/browserHostService';
-import { CommandsRegistry } from 'vs/platform/commands/common/commands';
+import { IUpdateProvider } from 'vs/workbench/services/update/browser/updateService';
+import { Event } from 'vs/base/common/event';
+import { IWorkspaceProvider } from 'vs/workbench/services/host/browser/browserHostService';
 import { IProductConfiguration } from 'vs/base/common/product';
-import { mark, PerformanceMark } from 'vs/base/common/performance';
 import { ICredentialsProvider } from 'vs/platform/credentials/common/credentials';
 import { TunnelProviderFeatures } from 'vs/platform/remote/common/tunnel';
-import { MenuId, MenuRegistry } from 'vs/platform/actions/common/actions';
-import { DeferredPromise } from 'vs/base/common/async';
-import { asArray } from 'vs/base/common/arrays';
 
-interface IResourceUriProvider {
+export interface IResourceUriProvider {
 	(uri: URI): URI;
 }
 
 /**
  * The identifier of an extension in the format: `PUBLISHER.NAME`. For example: `vscode.csharp`
  */
-type ExtensionId = string;
+export type ExtensionId = string;
 
-type MarketplaceExtension = ExtensionId | { readonly id: ExtensionId, preRelease?: boolean, migrateStorageFrom?: ExtensionId };
+export type MarketplaceExtension = ExtensionId | { readonly id: ExtensionId, preRelease?: boolean, migrateStorageFrom?: ExtensionId };
 
-interface ICommonTelemetryPropertiesResolver {
+export interface ICommonTelemetryPropertiesResolver {
 	(): { [key: string]: any };
 }
 
-interface IExternalUriResolver {
+export interface IExternalUriResolver {
 	(uri: URI): Promise<URI>;
 }
 
-interface IExternalURLOpener {
+export interface IExternalURLOpener {
 
 	/**
 	 * Overrides the behavior when an external URL is about to be opened.
@@ -52,7 +45,7 @@ interface IExternalURLOpener {
 	openExternal(href: string): boolean | Promise<boolean>;
 }
 
-interface ITunnelProvider {
+export interface ITunnelProvider {
 
 	/**
 	 * Support for creating tunnels.
@@ -70,11 +63,11 @@ interface ITunnelProvider {
 	features?: TunnelProviderFeatures;
 }
 
-interface ITunnelFactory {
+export interface ITunnelFactory {
 	(tunnelOptions: ITunnelOptions, tunnelCreationOptions: TunnelCreationOptions): Promise<ITunnel> | undefined;
 }
 
-interface ITunnelOptions {
+export interface ITunnelOptions {
 
 	remoteAddress: { port: number, host: string };
 
@@ -95,7 +88,7 @@ interface ITunnelOptions {
 	protocol?: string;
 }
 
-interface TunnelCreationOptions {
+export interface TunnelCreationOptions {
 
 	/**
 	 * True when the local operating system will require elevation to use the requested local port.
@@ -103,7 +96,7 @@ interface TunnelCreationOptions {
 	elevationRequired?: boolean;
 }
 
-interface ITunnel {
+export interface ITunnel {
 
 	remoteAddress: { port: number, host: string };
 
@@ -132,23 +125,16 @@ interface ITunnel {
 	dispose(): Promise<void> | void;
 }
 
-interface IShowPortCandidate {
+export interface IShowPortCandidate {
 	(host: string, port: number, detail: string): Promise<boolean>;
 }
 
-enum Menu {
+export enum Menu {
 	CommandPalette,
 	StatusBarWindowIndicatorMenu,
 }
 
-function asMenuId(menu: Menu): MenuId {
-	switch (menu) {
-		case Menu.CommandPalette: return MenuId.CommandPalette;
-		case Menu.StatusBarWindowIndicatorMenu: return MenuId.StatusBarWindowIndicatorMenu;
-	}
-}
-
-interface ICommand {
+export interface ICommand {
 
 	/**
 	 * An identifier for the command. Commands can be executed from extensions
@@ -179,7 +165,7 @@ interface ICommand {
 	handler: (...args: any[]) => unknown;
 }
 
-interface IHomeIndicator {
+export interface IHomeIndicator {
 
 	/**
 	 * The link to open when clicking the home indicator.
@@ -198,7 +184,7 @@ interface IHomeIndicator {
 	title: string;
 }
 
-interface IWelcomeBanner {
+export interface IWelcomeBanner {
 
 	/**
 	 * Welcome banner message to appear as text.
@@ -219,7 +205,7 @@ interface IWelcomeBanner {
 	actions?: IWelcomeBannerAction[];
 }
 
-interface IWelcomeBannerAction {
+export interface IWelcomeBannerAction {
 
 	/**
 	 * The link to open when clicking. Supports command invocation when
@@ -238,7 +224,7 @@ interface IWelcomeBannerAction {
 	title?: string;
 }
 
-interface IWindowIndicator {
+export interface IWindowIndicator {
 
 	/**
 	 * Triggering this event will cause the window indicator to update.
@@ -264,13 +250,13 @@ interface IWindowIndicator {
 	command?: string;
 }
 
-enum ColorScheme {
+export enum ColorScheme {
 	DARK = 'dark',
 	LIGHT = 'light',
 	HIGH_CONTRAST = 'hc'
 }
 
-interface IInitialColorTheme {
+export interface IInitialColorTheme {
 
 	/**
 	 * Initial color theme type.
@@ -283,16 +269,16 @@ interface IInitialColorTheme {
 	readonly colors?: { [colorId: string]: string };
 }
 
-interface IDefaultView {
+export interface IDefaultView {
 	readonly id: string;
 }
 
-interface IPosition {
+export interface IPosition {
 	readonly line: number;
 	readonly column: number;
 }
 
-interface IRange {
+export interface IRange {
 
 	/**
 	 * The start position. It is before or equal to end position.
@@ -305,14 +291,14 @@ interface IRange {
 	readonly end: IPosition;
 }
 
-interface IDefaultEditor {
+export interface IDefaultEditor {
 	readonly uri: UriComponents;
 	readonly selection?: IRange;
 	readonly openOnlyIfExists?: boolean;
 	readonly openWith?: string;
 }
 
-interface IDefaultLayout {
+export interface IDefaultLayout {
 	readonly views?: IDefaultView[];
 	readonly editors?: IDefaultEditor[];
 
@@ -323,7 +309,7 @@ interface IDefaultLayout {
 	readonly force?: boolean;
 }
 
-interface IProductQualityChangeHandler {
+export interface IProductQualityChangeHandler {
 
 	/**
 	 * Handler is being called when the user wants to switch between
@@ -335,7 +321,7 @@ interface IProductQualityChangeHandler {
 /**
  * Settings sync options
  */
-interface ISettingsSyncOptions {
+export interface ISettingsSyncOptions {
 
 	/**
 	 * Is settings sync enabled
@@ -354,7 +340,7 @@ interface ISettingsSyncOptions {
 	enablementHandler?(enablement: boolean): void;
 }
 
-interface IDevelopmentOptions {
+export interface IDevelopmentOptions {
 
 	/**
 	 * Current logging level. Default is `LogLevel.Info`.
@@ -377,7 +363,7 @@ interface IDevelopmentOptions {
 	readonly enableSmokeTestDriver?: boolean;
 }
 
-interface IWorkbenchConstructionOptions {
+export interface IWorkbenchConstructionOptions {
 
 	//#region Connection related configuration
 
@@ -572,211 +558,3 @@ interface IWorkbenchConstructionOptions {
 	//#endregion
 
 }
-
-/**
- * Creates the workbench with the provided options in the provided container.
- *
- * @param domElement the container to create the workbench in
- * @param options for setting up the workbench
- */
-let created = false;
-const workbenchPromise = new DeferredPromise<IWorkbench>();
-function create(domElement: HTMLElement, options: IWorkbenchConstructionOptions): IDisposable {
-
-	// Mark start of workbench
-	mark('code/didLoadWorkbenchMain');
-
-	// Assert that the workbench is not created more than once. We currently
-	// do not support this and require a full context switch to clean-up.
-	if (created) {
-		throw new Error('Unable to create the VSCode workbench more than once.');
-	} else {
-		created = true;
-	}
-
-	// Register commands if any
-	if (Array.isArray(options.commands)) {
-		for (const c of options.commands) {
-			const command: ICommand = c;
-
-			CommandsRegistry.registerCommand(command.id, (accessor, ...args) => {
-				// we currently only pass on the arguments but not the accessor
-				// to the command to reduce our exposure of internal API.
-				return command.handler(...args);
-			});
-
-			// Commands with labels appear in the command palette
-			if (command.label) {
-				for (const menu of asArray(command.menu ?? Menu.CommandPalette)) {
-					MenuRegistry.appendMenuItem(asMenuId(menu), { command: { id: command.id, title: command.label } });
-				}
-			}
-		}
-	}
-
-	CommandsRegistry.registerCommand('_workbench.getTarballProxyEndpoints', () => (options._tarballProxyEndpoints ?? {}));
-
-	// Startup workbench and resolve waiters
-	let instantiatedWorkbench: IWorkbench | undefined = undefined;
-	main(domElement, options).then(workbench => {
-		instantiatedWorkbench = workbench;
-		workbenchPromise.complete(workbench);
-	});
-
-	return toDisposable(() => {
-		if (instantiatedWorkbench) {
-			instantiatedWorkbench.shutdown();
-		} else {
-			workbenchPromise.p.then(instantiatedWorkbench => instantiatedWorkbench.shutdown());
-		}
-	});
-}
-
-
-//#region API Facade
-
-namespace commands {
-
-	/**
-	* Allows to execute any command if known with the provided arguments.
-	*
-	* @param command Identifier of the command to execute.
-	* @param rest Parameters passed to the command function.
-	* @return A promise that resolves to the returned value of the given command.
-	*/
-	export async function executeCommand(command: string, ...args: any[]): Promise<unknown> {
-		const workbench = await workbenchPromise.p;
-
-		return workbench.commands.executeCommand(command, ...args);
-	}
-}
-
-namespace env {
-
-	/**
-	 * Retrieve performance marks that have been collected during startup. This function
-	 * returns tuples of source and marks. A source is a dedicated context, like
-	 * the renderer or an extension host.
-	 *
-	 * *Note* that marks can be collected on different machines and in different processes
-	 * and that therefore "different clocks" are used. So, comparing `startTime`-properties
-	 * across contexts should be taken with a grain of salt.
-	 *
-	 * @returns A promise that resolves to tuples of source and marks.
-	 */
-	export async function retrievePerformanceMarks(): Promise<[string, readonly PerformanceMark[]][]> {
-		const workbench = await workbenchPromise.p;
-
-		return workbench.env.retrievePerformanceMarks();
-	}
-
-	/**
-	 * @returns the scheme to use for opening the associated desktop
-	 * experience via protocol handler.
-	 */
-	export async function getUriScheme(): Promise<string> {
-		const workbench = await workbenchPromise.p;
-
-		return workbench.env.uriScheme;
-	}
-
-	/**
-	 * Allows to open a `URI` with the standard opener service of the
-	 * workbench.
-	 */
-	export async function openUri(target: URI): Promise<boolean> {
-		const workbench = await workbenchPromise.p;
-
-		return workbench.env.openUri(target);
-	}
-}
-
-export {
-
-	// Factory
-	create,
-	IWorkbenchConstructionOptions,
-	IWorkbench,
-
-	// Basic Types
-	URI,
-	UriComponents,
-	Event,
-	Emitter,
-	IDisposable,
-	Disposable,
-
-	// Workspace
-	IWorkspace,
-	IWorkspaceProvider,
-
-	// WebSockets
-	IWebSocketFactory,
-	IWebSocket,
-
-	// Resources
-	IResourceUriProvider,
-
-	// Credentials
-	ICredentialsProvider,
-
-	// Callbacks
-	IURLCallbackProvider,
-
-	// LogLevel
-	LogLevel,
-
-	// SettingsSync
-	ISettingsSyncOptions,
-
-	// Updates/Quality
-	IUpdateProvider,
-	IUpdate,
-	IProductQualityChangeHandler,
-
-	// Telemetry
-	ICommonTelemetryPropertiesResolver,
-
-	// External Uris
-	IExternalUriResolver,
-
-	// External URL Opener
-	IExternalURLOpener,
-
-	// Tunnel
-	ITunnelProvider,
-	ITunnelFactory,
-	ITunnel,
-	ITunnelOptions,
-
-	// Ports
-	IShowPortCandidate,
-
-	// Commands
-	ICommand,
-	commands,
-	Menu,
-
-	// Branding
-	IHomeIndicator,
-	IWelcomeBanner,
-	IWelcomeBannerAction,
-	IProductConfiguration,
-	IWindowIndicator,
-	IInitialColorTheme,
-
-	// Default layout
-	IDefaultView,
-	IDefaultEditor,
-	IDefaultLayout,
-	IPosition,
-	IRange as ISelection,
-
-	// Env
-	env,
-
-	// Development
-	IDevelopmentOptions
-};
-
-//#endregion
