+++ src/vs/workbench/services/environment/common/environmentService.ts
@@ -8,7 +8,7 @@ import { IWindowConfiguration } from 'vs/platform/windows/common/windows';
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
 // TODO@bpasero layers
 // eslint-disable-next-line code-layering, code-import-patterns
-import type { IWorkbenchConstructionOptions as IWorkbenchOptions } from 'vs/workbench/browser/web.api';
+import { IWorkbenchConstructionOptions as IWorkbenchOptions } from 'vs/workbench/browser/web.api';
 import { URI } from 'vs/base/common/uri';
 
 export const IWorkbenchEnvironmentService = refineServiceDecorator<IEnvironmentService, IWorkbenchEnvironmentService>(IEnvironmentService);
